ux: keyboard shortcuts now do not need settings section to be focused to work
Until now, the keyboard shortcuts for selecting character types and for
generating sequences had only worked if the settings section had been in
focus. (Because of this, a hidden HTML element had to be focused on page
load via JS for the shortcuts to work on load. And this was being done in
a way that was getting a warning from the WAVE accessibility extension.)

Now the JS event-listener listens to 'document' so that the shortcuts can
be used freely. An additional if-condition was added to stop the shortcuts
from working when keypresses come from within input fields.